webapp: render pages optimistically instead of gating on the socket
The webapp is always robot-served, so the connection target is known
(location.hostname) and the address card is only a failure fallback — no reason
to hide the page until rosbridge connects. Each page is its own document, so
navigating reconnects from scratch; gating on "connected" made every hop show
the connect card until the socket came up — visibly slow over mDNS
(mars.local), where name resolution adds latency to each fresh connection.

Extract a shared mountPage(): build the view immediately, auto-connect to the
serving host in the background, and only show the connect card on a fail-fast
"disconnected" (a first connect that never opened, or laptop-dev idle). Panels
subscribe right away and fill in once connected (rosClient re-subscribes), so a
page switch paints instantly and the header badge carries the link status.
Replaces the duplicated connect/cockpit gate in all five page entries.
