+++ webapp/js/pageMount.js
@@ -1,0 +1,54 @@
+// @ts-check
+// Optimistic page mount, shared by every page entry.
+//
+// The robot always serves the webapp, so the connection target is known
+// (location.hostname) and the address-entry card is only a failure fallback —
+// there's no reason to gate the whole page on the socket. So: build the view
+// immediately and keep it on screen, auto-connect to the serving host in the
+// background, and let each panel show its own loading state until the socket is
+// up (rosClient re-subscribes on connect). A page switch paints instantly
+// instead of waiting on the (mDNS-resolved) reconnect; only a genuine
+// first-connect failure brings the connect card back.
+
+import { ros } from "./rosClient.js";
+import { createConnectPanel } from "./teleop/connectPanel.js";
+
+/**
+ * @param {HTMLElement} stage The page's #stage element.
+ * @param {string} viewClass CSS class for the view layer (e.g. "cockpit").
+ * @param {(root: HTMLElement) => { destroy: () => void }} buildView
+ *   Builds the page's content; called once, immediately, before the socket is up.
+ * @returns {{ destroy: () => void }} the built view
+ */
+export function mountPage(stage, viewClass, buildView) {
+  const connectLayer = document.createElement("div");
+  connectLayer.className = "connect-layer";
+  const viewLayer = document.createElement("div");
+  viewLayer.className = viewClass;
+  stage.append(connectLayer, viewLayer);
+
+  createConnectPanel(connectLayer, ros);
+
+  // Build now — don't wait for the socket. Panels subscribe right away and fill
+  // in once connected.
+  const view = buildView(viewLayer);
+
+  // Robot-served → reach the serving host immediately. Localhost (laptop dev)
+  // has no robot to auto-connect to, so it keeps the manual connect flow (the
+  // card stays up until the operator enters an address).
+  const servedHost = location.hostname;
+  if (servedHost && servedHost !== "localhost" && servedHost !== "127.0.0.1") {
+    ros.connect(servedHost);
+  }
+
+  // Keep the view up through connecting / connected / reconnecting — the header
+  // badge carries the link status. Only a fail-fast "disconnected" (a first
+  // connect that never opened, or the idle laptop-dev state) shows the card.
+  ros.onStateChange((state) => {
+    const failed = state === "disconnected";
+    connectLayer.hidden = !failed;
+    viewLayer.hidden = failed;
+  });
+
+  return view;
+}
