+++ webapp/js/teleop/main.js
@@ -12,7 +12,7 @@ import { ros } from "../rosClient.js";
 import { drive } from "../driveController.js";
 import { initShell } from "../shell.js";
 import { WebRtcSession } from "../webrtcSession.js";
-import { createConnectPanel } from "./connectPanel.js";
+import { mountPage } from "../pageMount.js";
 import { createVideoStage, createAudioToggle } from "./videoStage.js";
 import { createJoystick } from "./joystick.js";
 import { createKeyboardDrive, createWasdChips } from "./keyboardDrive.js";
@@ -30,37 +30,7 @@ const dbg = { ros, drive, session: null };
 
 const stage = /** @type {HTMLElement} */ (document.getElementById("stage"));
 
-const connectLayer = document.createElement("div");
-connectLayer.className = "connect-layer";
-const cockpitLayer = document.createElement("div");
-cockpitLayer.className = "cockpit";
-cockpitLayer.hidden = true;
-stage.append(connectLayer, cockpitLayer);
-
-createConnectPanel(connectLayer, ros);
-
-// Served by the robot itself → connect to it straight away; the connect
-// panel stays underneath as the fallback if that fails. Laptop-served
-// (localhost) keeps the manual flow.
-const servedHost = location.hostname;
-if (servedHost && servedHost !== "localhost" && servedHost !== "127.0.0.1") {
-  ros.connect(servedHost);
-}
-
-/** @type {{ destroy: () => void } | null} */
-let cockpit = null;
-
-ros.onStateChange((state) => {
-  const showCockpit = state === "connected" || (state === "reconnecting" && cockpit !== null);
-  connectLayer.hidden = showCockpit;
-  cockpitLayer.hidden = !showCockpit;
-  if (state === "connected" && !cockpit) {
-    cockpit = buildCockpit(cockpitLayer);
-  } else if (state === "disconnected" && cockpit) {
-    cockpit.destroy();
-    cockpit = null;
-  }
-});
+mountPage(stage, "cockpit", buildCockpit);
 
 /**
  * @param {HTMLElement} root
